+++ XamoomSDK/Classes/mapping/FlowLayout.m
@@ -1,0 +1,16 @@
+//
+//  FlowLayout.m
+//  XamoomSDK
+//
+//  Created by Thomas Krainz-Mischitz on 31.07.19.
+//
+
+#import "FlowLayout.h"
+
+@implementation FlowLayout
+
+- (BOOL)shouldInvalidateLayoutForBoundsChange:(CGRect)newBounds {
+  return YES;
+}
+
+@end